Rename FeaturedProductSenseiPro to SenseiProAd
